refactor: simplify options type (#7420)
