+++ packages/better-auth/src/adapters/tests/adapter-factory.test.ts
@@ -7,7 +7,7 @@ import type {
 } from "@better-auth/core/db/adapter";
 import { createAdapterFactory } from "@better-auth/core/db/adapter";
 import { describe, expect, test } from "vitest";
-import { betterAuth } from "../../auth";
+import { betterAuth } from "../../auth/full";
 import type { User } from "../../types";
 
 /*
