+++ packages/better-auth/src/auth/base.ts
@@ -7,9 +7,7 @@ import type { Auth } from "../types";
 import { getBaseURL, getOrigin } from "../utils/url";
 
 export const createBetterAuth = <Options extends BetterAuthOptions>(
-	options: Options &
-		// fixme(alex): do we need Record<never, never> here?
-		Record<never, never>,
+	options: Options,
 	initFn: (options: Options) => Promise<AuthContext>,
 ): Auth<Options> => {
 	const authContext = initFn(options);
