+++ packages/better-auth/src/db/internal-adapter.test.ts
@@ -3,7 +3,7 @@ import { safeJSONParse } from "@better-auth/core/utils";
 import Database from "better-sqlite3";
 import { Kysely, SqliteDialect } from "kysely";
 import { afterEach, beforeAll, describe, expect, it, vi } from "vitest";
-import { betterAuth } from "../auth";
+import { betterAuth } from "../auth/full";
 import { init } from "../context/init";
 import { getTestInstance } from "../test-utils/test-instance";
 import type {
